Pacto personified
I rewrote the intro to Pacto with a central metaphor/personification: Pacto a a judge who arbitrates contract disputes.

I don't want to fall into a trap of writing "legalese" instead of simple English, but I think borrowing some real-world metaphors and terms like **clauses** or **breach of contract** could help clarify the components of Pacto - both for users and contributors.  I think we've spent too much time mixing up an actual request with a **request clause**, which describes the obligations a request must meet, but is not a valid request itself.

It's a bit long, and could use some images (I've been creating some for the site), but what do you think?